+++ studio/src/app/events/editor/doc/doc.data.events.ts
@@ -1,13 +1,11 @@
-import {Doc, DocData, now, Paragraph, ParagraphData, elementIndex, isTextNode, cleanNode, isElementNode} from '@deckdeckgo/editor';
+import {v4 as uuid} from 'uuid';
 
-import {debounce} from '@deckdeckgo/utils';
+import {Doc, DocData, now, Paragraph, ParagraphData, elementIndex, isTextNode, cleanNode, isElementNode} from '@deckdeckgo/editor';
 
 import errorStore from '../../../stores/error.store';
 import busyStore from '../../../stores/busy.store';
 import editorStore from '../../../stores/editor.store';
 import authStore from '../../../stores/auth.store';
-
-import {findParagraph} from '../../../utils/editor/paragraph.utils';
 
 import {createOfflineDoc, updateOfflineDoc} from '../../../providers/data/doc/doc.offline.provider';
 import {
@@ -15,88 +13,66 @@ import {
   deleteOfflineParagraph,
   updateOfflineParagraph
 } from '../../../providers/data/paragraph/paragraph.offline.provider';
-import {
-  findAddedNodesParagraphs,
-  findAddedParagraphs,
-  findRemovedNodesParagraphs,
-  findUpdatedParagraphs
-} from '../../../utils/editor/paragraphs.utils';
 
 export class DocDataEvents {
-  private containerRef: HTMLElement;
-
-  private treeObserver: MutationObserver | undefined;
-  private attributesObserver: MutationObserver | undefined;
-  private dataObserver: MutationObserver | undefined;
-
-  private stackDataMutations: MutationRecord[] = [];
-
-  private readonly debounceUpdateInput: () => void = debounce(async () => await this.updateData(), 500);
-
-  private debounceBusyEnd: () => void = debounce(() => (busyStore.state.busy = false), 500);
-
-  init(containerRef: HTMLElement) {
-    this.containerRef = containerRef;
-
-    this.treeObserver = new MutationObserver(this.onTreeMutation);
-    this.treeObserver.observe(this.containerRef, {childList: true, subtree: true});
-
-    this.attributesObserver = new MutationObserver(this.onAttributesMutation);
-    this.attributesObserver.observe(this.containerRef, {attributes: true, subtree: true});
-
-    this.dataObserver = new MutationObserver(this.onDataMutation);
-    this.dataObserver.observe(this.containerRef, {characterData: true, subtree: true});
-
-    document.addEventListener('markdownDidChange', this.onCustomEventChange);
-    document.addEventListener('wordCloudDidChange', this.onCustomEventChange);
-    document.addEventListener('codeDidChange', this.onCustomEventChange);
-    document.addEventListener('mathDidChange', this.onCustomEventChange);
-    document.addEventListener('imgDidChange', this.onCustomEventChange);
+  init() {
+    document.addEventListener('addParagraphs', this.onAddParagraphs);
+    document.addEventListener('deleteParagraphs', this.onDeleteParagraphs);
+    document.addEventListener('updateParagraphs', this.onUpdateParagraphs);
   }
 
   destroy() {
-    this.treeObserver?.disconnect();
-    this.attributesObserver?.disconnect();
-    this.dataObserver?.disconnect();
-
-    document.addEventListener('markdownDidChange', this.onCustomEventChange);
-    document.addEventListener('wordCloudDidChange', this.onCustomEventChange);
-    document.addEventListener('codeDidChange', this.onCustomEventChange);
-    document.addEventListener('mathDidChange', this.onCustomEventChange);
-    document.addEventListener('imgDidChange', this.onCustomEventChange);
+    document.removeEventListener('addParagraphs', this.onAddParagraphs);
+    document.removeEventListener('deleteParagraphs', this.onDeleteParagraphs);
+    document.removeEventListener('updateParagraphs', this.onUpdateParagraphs);
   }
 
-  private onTreeMutation = async (mutations: MutationRecord[]) => {
-    busyStore.state.busy = true;
+  private onAddParagraphs = async ({detail: addedParagraphs}: CustomEvent<HTMLElement[]>) => {
+    try {
+      busyStore.state.busy = true;
 
-    await this.addParagraphs(mutations);
-    await this.deleteParagraphs(mutations);
-    await this.updateAddedNodesParagraphs(mutations);
+      await this.createDoc();
 
-    this.debounceBusyEnd();
-  };
-
-  private onAttributesMutation = async (mutations: MutationRecord[]) => {
-    busyStore.state.busy = true;
-
-    await this.updateParagraphs(mutations.filter(({attributeName}: MutationRecord) => ['style'].includes(attributeName)));
-
-    this.debounceBusyEnd();
-  };
-
-  private onDataMutation = (mutations: MutationRecord[]) => {
-    this.stackDataMutations.push(...mutations);
-    this.debounceUpdateInput();
-  };
-
-  private onCustomEventChange = async ({detail}: CustomEvent<HTMLElement>) => {
-    const paragraph: Node | undefined = findParagraph({element: detail, container: this.containerRef});
-
-    if (!paragraph || isTextNode(paragraph)) {
-      return;
+      for (const paragraph of addedParagraphs) {
+        await this.createParagraph(paragraph);
+      }
+    } catch (err) {
+      errorStore.state.error = err;
     }
 
-    await this.updateParagraph(paragraph as HTMLElement);
+    busyStore.state.busy = false;
+  };
+
+  private onDeleteParagraphs = async ({detail: removedParagraphs}: CustomEvent<HTMLElement[]>) => {
+    try {
+      busyStore.state.busy = true;
+
+      const promises: Promise<string | undefined>[] = removedParagraphs.map((paragraph: HTMLElement) => this.deleteParagraph(paragraph));
+      const removedParagraphIds: (string | undefined)[] = await Promise.all(promises);
+
+      await this.filterDocParagraphList(removedParagraphIds);
+    } catch (err) {
+      errorStore.state.error = err;
+    }
+
+    busyStore.state.busy = false;
+  };
+
+  private onUpdateParagraphs = async ({detail: updatedParagraphs}: CustomEvent<HTMLElement[]>) => {
+    try {
+      busyStore.state.busy = true;
+
+      // In case of copy-paste, the browser might proceed with a delete-update for which we do not get paragraph_id.
+      // It might copy a <p/> within a <div/> instead of creating a child of the container for the new <p/>
+      const promises: Promise<void>[] = updatedParagraphs
+        .filter((paragraph: HTMLElement) => paragraph.hasAttribute('paragraph_id'))
+        .map((paragraph: HTMLElement) => this.updateParagraph(paragraph));
+      await Promise.all(promises);
+    } catch (err) {
+      errorStore.state.error = err;
+    }
+
+    busyStore.state.busy = false;
   };
 
   private createDoc(): Promise<void> {
@@ -127,13 +103,10 @@ export class DocDataEvents {
     await this.updateDocParagraphList({paragraphId, paragraphElement: element});
   }
 
-  private async deleteParagraph(element: Node): Promise<string | undefined> {
-    if (element.nodeType === Node.TEXT_NODE || element.nodeType === Node.COMMENT_NODE) {
-      return;
-    }
+  private async deleteParagraph(element: HTMLElement): Promise<string | undefined> {
+    const paragraphId: string = element.getAttribute('paragraph_id');
 
-    const paragraphId: string = (element as HTMLElement).getAttribute('paragraph_id');
-
+    // Cannot happen because paragraph_id are filter in stylo before triggering the delete paragraphs event
     if (!paragraphId) {
       return undefined;
     }
@@ -159,11 +132,15 @@ export class DocDataEvents {
         paragraphData.children = content;
       }
 
-      const persistedParagraph: Paragraph = await createOfflineParagraph({docId: editorStore.state.doc.id, paragraphData: paragraphData});
+      const paragraphId: string = uuid();
 
-      if (persistedParagraph && persistedParagraph.id) {
-        element.setAttribute('paragraph_id', persistedParagraph.id);
-      }
+      const persistedParagraph: Paragraph = await createOfflineParagraph({
+        docId: editorStore.state.doc.id,
+        paragraphData: paragraphData,
+        paragraphId
+      });
+
+      element.setAttribute('paragraph_id', paragraphId);
 
       resolve(persistedParagraph);
     });
@@ -235,101 +212,6 @@ export class DocDataEvents {
     });
   }
 
-  private async addParagraphs(mutations: MutationRecord[]) {
-    try {
-      if (!this.containerRef) {
-        return;
-      }
-
-      const addedParagraphs: HTMLElement[] = findAddedParagraphs({mutations, container: this.containerRef});
-
-      await this.createDoc();
-
-      for (const paragraph of addedParagraphs) {
-        await this.createParagraph(paragraph);
-      }
-    } catch (err) {
-      errorStore.state.error = err;
-    }
-  }
-
-  private async deleteParagraphs(mutations: MutationRecord[]) {
-    try {
-      if (!this.containerRef) {
-        return;
-      }
-
-      if (!mutations || mutations.length <= 0) {
-        return;
-      }
-
-      const removedNodes: Node[] = mutations.reduce(
-        (acc: Node[], {removedNodes}: MutationRecord) => [...acc, ...Array.from(removedNodes)],
-        []
-      );
-
-      const promises: Promise<string | undefined>[] = removedNodes.map((node: Node) => this.deleteParagraph(node));
-      const removedParagraphIds: (string | undefined)[] = await Promise.all(promises);
-
-      await this.filterDocParagraphList(removedParagraphIds);
-    } catch (err) {
-      errorStore.state.error = err;
-    }
-  }
-
-  private async updateAddedNodesParagraphs(mutations: MutationRecord[]) {
-    try {
-      if (!this.containerRef) {
-        return;
-      }
-
-      if (!mutations || mutations.length <= 0) {
-        return;
-      }
-
-      const addedNodesMutations: MutationRecord[] = findAddedNodesParagraphs({mutations, container: this.containerRef});
-      const removedNodesMutations: MutationRecord[] = findRemovedNodesParagraphs({mutations, container: this.containerRef});
-
-      await this.updateParagraphs([...addedNodesMutations, ...removedNodesMutations]);
-    } catch (err) {
-      errorStore.state.error = err;
-    }
-  }
-
-  private async updateData() {
-    if (!this.stackDataMutations || this.stackDataMutations.length <= 0) {
-      return;
-    }
-
-    busyStore.state.busy = true;
-
-    const mutations: MutationRecord[] = [...this.stackDataMutations];
-    this.stackDataMutations = [];
-
-    await this.updateParagraphs(mutations);
-
-    this.debounceBusyEnd();
-  }
-
-  private async updateParagraphs(mutations: MutationRecord[]) {
-    try {
-      if (!this.containerRef) {
-        return;
-      }
-
-      const updateParagraphs: HTMLElement[] = findUpdatedParagraphs({mutations, container: this.containerRef});
-
-      if (updateParagraphs.length <= 0) {
-        return;
-      }
-
-      const promises: Promise<void>[] = updateParagraphs.map((paragraph: HTMLElement) => this.updateParagraph(paragraph));
-      await Promise.all(promises);
-    } catch (err) {
-      errorStore.state.error = err;
-    }
-  }
-
   private async updateParagraph(paragraph: HTMLElement) {
     if (!paragraph || !paragraph.nodeName) {
       return;
@@ -342,7 +224,7 @@ export class DocDataEvents {
       return;
     }
 
-    const paragraphId: string = paragraph.getAttribute('paragraph_id');
+    const paragraphId: string | null = paragraph.getAttribute('paragraph_id');
 
     if (!paragraphId) {
       errorStore.state.error = 'Paragraph is not defined';